ENG-163: metrics wrapper class and (mostly) transparent metrics collection for storage, deduplicator, and directory service (#273)

+++ testing/start-e2e.sh
@@ -78,11 +78,17 @@ fi
 export UH_LICENSE="$(cat $PWD/../data/licenses/UltiHash-Test-1GB.lic)"
 
 if [ -z "$cluster_url" ]; then
+    if ! docker pull ghcr.io/ultihash/build-base:latest; then
+        echo "pulling build-base image failed" 1>&2
+        exit 1
+    fi
+
+    docker image rm uh-cluster:testing
+
     if ! docker build --no-cache --file ../Dockerfile --tag uh-cluster:testing ..; then
         echo "docker build failed" 1>&2
         exit 1
     fi
-
     docker compose up --detach
     trap "docker-compose down" SIGHUP SIGINT SIGQUIT SIGABRT EXIT
     cluster_url="http://localhost:8080"
